Change all fields and database rows with ID of entities from type int to type long. Also fix tests

+++ src/main/java/com/fatykhov/restapp/dto/ActorMovieDto.java
@@ -16,6 +16,6 @@ import java.util.List;
 @Setter
 @ToString
 public class ActorMovieDto {
-    private List<Integer> actorsId;
+    private List<Long> actorsId;
     private MovieDto movieDto;
 }
